format

+++ mobile/lib/Screens/Shared/Models/review_model.dart
@@ -14,7 +14,6 @@ class ReviewModel {
     required this.stars,
     required this.description,
   });
-  
 
   ReviewModel copyWith({
     int? id,
@@ -47,14 +46,15 @@ class ReviewModel {
       id: map['id'],
       firstName: map['first_name'],
       lastName: map['last_name'],
-      stars: map['stars'],
+      stars: map['stars'].toDouble(),
       description: map['description'],
     );
   }
 
   String toJson() => json.encode(toMap());
 
-  factory ReviewModel.fromJson(String source) => ReviewModel.fromMap(json.decode(source));
+  factory ReviewModel.fromJson(String source) =>
+      ReviewModel.fromMap(json.decode(source));
 
   @override
   String toString() {
@@ -64,21 +64,21 @@ class ReviewModel {
   @override
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
-  
+
     return other is ReviewModel &&
-      other.id == id &&
-      other.firstName == firstName &&
-      other.lastName == lastName &&
-      other.stars == stars &&
-      other.description == description;
+        other.id == id &&
+        other.firstName == firstName &&
+        other.lastName == lastName &&
+        other.stars == stars &&
+        other.description == description;
   }
 
   @override
   int get hashCode {
     return id.hashCode ^
-      firstName.hashCode ^
-      lastName.hashCode ^
-      stars.hashCode ^
-      description.hashCode;
+        firstName.hashCode ^
+        lastName.hashCode ^
+        stars.hashCode ^
+        description.hashCode;
   }
 }
